Refactor studio-dashboard: split 1,971 lines into 6 specialized modules
- Extract dashboard-state.js (137 lines) - data loading & real-time updates
- Extract dashboard-ui.js (150 lines) - notifications, toasts, UI utilities
- Extract dashboard-agents.js (169 lines) - agent list & management
- Extract dashboard-renderer.js (367 lines) - main render, metrics, progress
- Extract dashboard-tasks.js (446 lines) - task list, CRUD, details modal
- Extract dashboard-scheduled.js (468 lines) - scheduled tasks management
- Main dashboard now orchestrator (143 lines, 92.7% reduction)
- All modules under 500 lines ✓
